+++ src/collective/compoundcriterion/vocabularies.py
@@ -1,16 +1,16 @@
-from zope.component import getGlobalSiteManager
-from zope.interface import implements
-from zope.schema.interfaces import IVocabularyFactory
-from zope.schema.vocabulary import SimpleVocabulary
-from zope.schema.vocabulary import SimpleTerm
-from collective.compoundcriterion.interfaces import ICompoundCriterionFilter
 from Products.CMFPlone import PloneMessageFactory as _p
+from collective.compoundcriterion.interfaces import ICompoundCriterionFilter
+from zope.component import getGlobalSiteManager
+from zope.interface import implementer
+from zope.schema.interfaces import IVocabularyFactory
+from zope.schema.vocabulary import SimpleTerm
+from zope.schema.vocabulary import SimpleVocabulary
 
 
+@implementer(IVocabularyFactory)
 class CompoundCriterionVocabulary(object):
     """Vocabulary factory for compound criterion.
        This will return every named adapter that provides the ICoumpondCriterionFilter interface."""
-    implements(IVocabularyFactory)
 
     def __call__(self, context):
         sm = getGlobalSiteManager()